Added new HTML form template
A new HTML form has been added to the project. This includes a styled container for the form, input fields for name, subject and grade, as well as a submit button. The page also features a top navigation bar with links to 'Form' and 'Grades'. Styling has been applied throughout for improved user experience.